fix: 整数オーバーフロー修正・メモリ削減・カラーマップ統一・dem10b URL修正
- src/analyze.c, src/unionfind.h/c: int32_t ピクセルインデックスを uint32_t/size_t に変更
  （5239: 70146×38402px=2.7億 で int32_t 上限超えによる SIGSEGV を修正）
- src/analyze.c: Pixel 構造体(12B×N) → uint32_t インデックス配列(4B×N) に変更
  （qsort_r 使用のため _GNU_SOURCE 定義追加、メモリ約37%削減）
- src/elevation.c: elev_fill_nodata_dem10b() のインデックス計算を size_t に変更
- src/mesh_analyze.c: 標高カラーマップを terrain_viz.c 配色に統一
  （海面/NODATA は濃紺固定、陸地との視認性を改善）
- scripts/prefetch_tiles.py: dem10b URL を dem10b_png → dem_png に修正
- analysis/terrain_viz.c: カラー変換サンプルコードを tests/ から移動

+++ src/elevation.c
@@ -352,7 +352,7 @@ void elev_fill_nodata_dem10b(ElevTile *big, const char *base,
             int has_nodata = 0;
             for (int py = py_lo; py <= py_hi && !has_nodata; py++)
                 for (int px = px_lo; px <= px_hi && !has_nodata; px++)
-                    if (big->data[py * (int)big->width + px] == ELEV_NODATA)
+                    if (big->data[(size_t)py * big->width + px] == ELEV_NODATA)
                         has_nodata = 1;
             if (!has_nodata) continue;
 
@@ -367,7 +367,7 @@ void elev_fill_nodata_dem10b(ElevTile *big, const char *base,
 
             for (int py = py_lo; py <= py_hi; py++) {
                 for (int px = px_lo; px <= px_hi; px++) {
-                    int idx = py * (int)big->width + px;
+                    size_t idx = (size_t)py * big->width + px;
                     if (big->data[idx] != ELEV_NODATA) continue;
 
                     int z15_tx = range_x_min + (px - 1) / TILE_PIX;
